справления по второму экспертному ревью (P0/P1)
P0-1: I-type ANDI/ORI/XORI/SHLI/SHRI теперь используют PhysOp1 как source/destination (а не неявный R0)

P0-2: T3Interpreter понижен до Experimental (scalar subset) — arrays/structs/globals/string/float не реализованы

P0-3: bst.t очищен от XML-артефактов; tstring.asm урезан до валидного crt0

P0-4: Word18 << использует long intermediate (было int, overflow при shift>=3)

P0-5: обавлены LICENSE (MIT), .editorconfig, global.json

P1-1: Tryte.FromLong проверяет диапазон до cast к short

P1-2: DeviceManager.Read/Write выбрасывают InvalidOperationException для незарегистрированных портов

P1-3: dotnet.yml синхронизирован с build-and-test.yml (добавлен T3Simulator.Common.Tests)

+++ src/TritTypes/Word18.cs
@@ -112,11 +112,11 @@ namespace TritTypes
         public Word18 Negate() => -this;
 
         // Shifts
-        public static Word18 operator <<(Word18 t, int shift)
-        {
-            if (shift < 0) throw new ArgumentOutOfRangeException(nameof(shift));
-            return FromWrapped(t._value * Pow3(shift));
-        }
+    public static Word18 operator <<(Word18 t, int shift)
+    {
+        if (shift < 0) throw new ArgumentOutOfRangeException(nameof(shift));
+        return FromWrapped((long)t._value * Pow3(shift));
+    }
         public static Word18 operator >>(Word18 t, int shift)
         {
             if (shift < 0) throw new ArgumentOutOfRangeException(nameof(shift));
